+++ libs/filamentapp/src/FilamentApp.cpp
@@ -522,8 +522,6 @@ FilamentApp::Window::Window(FilamentApp* filamentApp,
         windowFlags |= SDL_WINDOW_HIDDEN;
     }
 
-    mBackend = config.backend;
-
     // Even if we're in headless mode, we still need to create a window, otherwise SDL will not poll
     // events.
     mWindow = SDL_CreateWindow(title.c_str(), x, y, (int) w, (int) h, windowFlags);
@@ -538,6 +536,9 @@ FilamentApp::Window::Window(FilamentApp* filamentApp,
         // For single-threaded platforms, we need to ensure that Filament's OpenGL context is
         // current, rather than the one created by SDL.
         mFilamentApp->mEngine = Engine::create(config.backend);
+
+        // get the resolved backend
+        mBackend = config.backend = mFilamentApp->mEngine->getBackend();
 
         void* nativeWindow = ::getNativeWindow(mWindow);
         void* nativeSwapChain = nativeWindow;